+++ CHANGELOG.md
@@ -31,8 +31,9 @@
 - Error handling - Make the zhinst backend correctly throw a `NotImplementedError` for `SetClockFrequency`. (!910)
 - Type checking - Add `isinstance` checks to help pyright. (!914)
 - Qblox backend - Remove pulsars from codebase. (!893)
-- tests - small refactor of the `mock_setup_basic_transmon` fixture. (!917)
-
+- tests 
+  - small refactor of the `mock_setup_basic_transmon` fixture. (!917)
+  - Cleaning unused `instrument.close()`. (!919)
 ## 0.18.0 (2023-12-22)
 
 ### Release highlights